refactor: use abstract factory for PDF and Excel export
