+++ salary-service/src/main/java/com/micro/salaryservice/service/ExportService.java
@@ -1,11 +1,11 @@
 package com.micro.salaryservice.service;
 
-import jakarta.servlet.http.HttpServletResponse;
+
 
 import java.io.IOException;
+import java.io.OutputStream;
 
 public interface ExportService {
-    Object exportExcelFile(HttpServletResponse response);
-
-    void exportPdfFile(HttpServletResponse response) throws IOException;
+    void export(OutputStream outputStream);
+    String getType();
 }
